@@ -16,8 +16,8 @@
     "axios": "^1.2.1",
     "degrees-to-direction": "^1.1.1",
     "framer-motion": "^4.1.17",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-icons": "^4.7.1",
     "react-router-dom": "^6.4.5",
     "react-scripts": "4.0.1",
