@@ -2,10 +2,15 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
+import { BrowserRouter } from 'react-router-dom';
+import { createRoot } from 'react-dom/client';
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
-);
+
+const domNode = document.getElementById('root');
+const root = createRoot(domNode);
+
+root.render( <React.StrictMode>
+  <BrowserRouter>
+  <App />
+  </BrowserRouter>
+</React.StrictMode>);
