@@ -1,60 +1,56 @@
 import { useEffect, useState } from 'react';
 import getWeather, { getForecast } from "./adapters/openweathermap.adapter"
-import  Layout from './components/Layout';
-import  City from './components/City';
-import  FutureWeathers from './components/FutureWeathers';
-import { Wrapper, Main, FormStyle, Forecast } from './style/weatherApp.style';
+import City from './components/City';
+import FutureWeathers from './components/FutureWeathers';
+import Layout from './components/Layout';
+import Search from './components/Search';
+import { Wrapper, Main, WeatherCard, Forecast } from './style/weatherApp.style';
 import fortawesome from '@fortawesome/fontawesome';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faSquarePlus, faSquareMinus } from '@fortawesome/free-solid-svg-icons';
-import { FaSearch } from "react-icons/fa"
+import { BrowserRouter, Router } from 'react-router-dom';
 
 
 
 function App() {
-
-  const [weatherData, setWeatherData] = useState([{}]);
-  const [input, setInput] = useState("");
+  
+  const [ weatherData, setWeatherData] = useState({loaded:false});
+  const [toggle, setToggle] = useState(false);
+  const [ forecastData, setForecastData] = useState({loaded:false});
 
   fortawesome.library.add(faSquarePlus, faSquareMinus);
 
-  
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    getWeather(input).then(dataWeather=>setWeatherData([{loaded:true, dataWeather, toggleForecast: false}])); 
-  }
+  useEffect(() => {
+    getWeather().then(data => setWeatherData({
+      loaded:true,
+      weather:data.weather[0],
+      main:data.main,
+      name:data.name,
+      wind:data.wind
+    }));
+  },[]);
 
-  const test = () => {
-    console.log("click");
-  }
-  console.log("data from forecast2", weatherData[0]);
- 
+  useEffect(() => {
+    
+    if (toggle === true) 
+      getForecast().then(data => setForecastData({ loaded:true, ...data}));
+  },[toggle]);
+  
   return (
-    <Wrapper>
-      <Layout>
-        <Main>
-        <FormStyle onSubmit={handleSubmit}> 
-          <div>
-              <FaSearch/>
-              <input onChange={(e)=> setInput(e.target.value)} type="text" value={input} />
-          </div>
-        </FormStyle>
-        {weatherData[0].loaded && (
-          <>
-            <City {...weatherData[0].dataWeather}/>
-            <FontAwesomeIcon onClick={()=>{ weatherData[0].toggleForecast = true; console.log("ulaz", this.weatherData[0].toggleForecast );}} icon={ weatherData[0].toggleForecast === false || undefined ? faSquarePlus : faSquareMinus} />
-          </>
-        )}
-        {weatherData[0].toggleForecast && (
-          <>
-          <Forecast>
-              <FutureWeathers { ...weatherData[0].dataForecast }/>
-          </Forecast>
-          </>
-        )}       
-        </Main>
-      </Layout>
-    </Wrapper>
+      <Wrapper>
+        <Layout>
+          <Main>
+            <Search></Search>
+            {weatherData.loaded && (<City {...weatherData}/>)}
+            <FontAwesomeIcon onClick={() => setToggle(!toggle)}  icon={ toggle === false || null ? faSquarePlus : faSquareMinus }/>
+            {forecastData.loaded && (<WeatherCard>
+            <Forecast className={toggle === false ? "hidden" : "show"}>
+              <FutureWeathers className={"weather"}{ ...forecastData }/>
+            </Forecast>
+          </WeatherCard>)}
+          </Main>
+        </Layout>
+      </Wrapper>
   );
 }
 
