@@ -1,25 +1,32 @@
+import { Main } from "../style/weatherApp.style"
+
 function City(props) {
   const d2d = require('degrees-to-direction');
+  console.log("props", props)
   const windDirection = d2d(props.wind.deg);
-  const iconSrc = `http://openweathermap.org/img/wn/${props.weather[0].icon}.png`;
- 
+
+  const iconSrc = `http://openweathermap.org/img/wn/${props.weather.icon}.png`;
+
   return (
     <>
+     <Main>
       <div className="city-info">
-        <p> {props.name}, {props.weather[0].description}</p>
+        <p> {props.name}, {props.weather.description}
+        </p>
         <img src={iconSrc} width="50px" height="50px" alt=""></img>
         </div>
         <div className="weather-info">
           <h2>
-            {(props.main.temp - 273.15).toFixed(0)}<span>&#xb0;</span>
+            {(props.main.temp - 273.15).toFixed(2)}<span>&#xb0;</span>
           </h2>
           <ul>
-            <li>Feels like: {(props.main.feels_like - 273.15).toFixed(0)} <span>&#8451;</span></li>
+            <li>Feels like: {(props.main.feels_like - 273.15).toFixed(2)} <span>&#8451;</span></li>
             <li>Pressure: {props.main.pressure} hPa</li>
             <li>Humidity: {props.main.humidity}%</li>
-            <li>Wind: {props.wind.speed.toFixed(0)} km/h, {windDirection} </li>
+            <li>Wind: {props.wind.speed} km/h, {windDirection} </li>
           </ul>
-        </div>
+          </div>
+        </Main>
     </>
   )
 }
